re-factor vector layers
