+++ CHANGES.txt
@@ -1,13 +1,14 @@
 0.6.0
 ~~~~~
 
-- Added `tooltip` support to `Marker`s (ocefpaf #722)
-- Added `tooltip` support to all vector layers (ocefpaf #721)
+- Internal re-factor to reflect leaflet's organization (ocefpaf #725)
+- Added `tooltip` support to `Marker`s (ocefpaf #724)
+- Added `tooltip` support to all vector layers (ocefpaf #722)
 
 API changes
 
 - `Rectangle` and `Polygon` were renamed and set to leaflet's defaults.
-  Both now accepted all `Path`'s optional arguments (ocefpaf #721)
+  Both now accepted all `Path`'s optional arguments (ocefpaf #722)
 
 Bug Fixes
