+++ folium/__init__.py
@@ -9,8 +9,8 @@ from branca.element import (CssLink, Div, Element, Figure, Html, IFrame,
 from folium._version import get_versions
 
 from folium.features import (
-    Circle, CircleMarker, ClickForMarker, CustomIcon, DivIcon, GeoJson,
-    LatLngPopup, PolyLine, RegularPolygonMarker, TopoJson, Vega, VegaLite,
+    ClickForMarker, ColorLine, CustomIcon, DivIcon, GeoJson,
+    LatLngPopup, RegularPolygonMarker, TopoJson, Vega, VegaLite,
     WmsTileLayer,
 )
 
@@ -19,6 +19,8 @@ from folium.folium import Map
 from folium.map import (
     FeatureGroup, FitBounds, Icon, LayerControl, Marker, Popup, TileLayer
 )
+
+from folium.vector_layers import Circle, CircleMarker, PolyLine, Polygon, Rectangle  # noqa
 
 __version__ = get_versions()['version']
 del get_versions
@@ -34,13 +36,10 @@ __all__ = [
     'Link',
     'MacroElement',
     'ColorMap',
+    'ColorLine',
     'LinearColormap',
     'StepColormap',
     'Map',
-    'Circle',
-    'CircleMarker',
-    'RectangleMarker',
-    'Polygon',
     'FeatureGroup',
     'FitBounds',
     'Icon',
@@ -55,10 +54,16 @@ __all__ = [
     'GeoJsonStyle',
     'LatLngPopup',
     'MarkerCluster',
-    'PolyLine',
     'Vega',
     'VegaLite',
     'RegularPolygonMarker',
     'TopoJson',
-    'WmsTileLayer'
+    'WmsTileLayer',
+    # vector_layers
+    'Circle',
+    'CircleMarker',
+    'PolyLine',
+    'Polygon',
+    'Polyline',
+    'Rectangle',
 ]
